refactor: 重構 app.js
