+++ routes/posts.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const PostController = require('../controller/post');
+const PostController = require('../controller/post_controller');
 const { isAuth } = require('../middleware/auth')
 
 module.exports = router;
